requirements.txt
aiocontextvars==0.2.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
ConfigArgParse==1.4.1
confluent-kafka==1.7.0
contextvars==2.4
dataclasses==0.8
Deprecated==1.2.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
google-api-core==1.29.0
google-auth==1.30.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
idna==2.10
immutables==0.15
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
locust==1.5.3
locust-plugins==1.1.0
lxml==4.6.3
MarkupSafe==2.0.1
msgpack==1.0.2
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
opentelemetry-api==1.3.0
packaging==20.9
protobuf==3.17.2
psutil==5.8.0
psycogreen==1.0.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
rsa==4.7.2
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
websocket-client==1.0.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
